+++ lib/rodauth/rails/app.rb
@@ -12,21 +12,15 @@ module Rodauth
       plugin :hooks
       plugin :render, layout: false
 
-      def self.configure(name = nil, **options, &block)
-        unless options[:json] == :only
-          require "rodauth/rails/app/flash"
-          plugin Flash
-        end
+      if defined?(ActionDispatch::Flash) # not in API-only mode
+        require "rodauth/rails/app/flash"
+        plugin Flash
+      end
 
-        plugin :rodauth, name: name, csrf: false, flash: false, **options do
+      def self.configure(name = nil, **options, &block)
+        plugin :rodauth, name: name, csrf: false, flash: false, json: true, **options do
           # load the Rails integration
           enable :rails
-
-          if options[:json] == :only && ActionPack.version >= Gem::Version.new("5.0")
-            rails_controller { ActionController::API }
-          else
-            rails_controller { ActionController::Base }
-          end
 
           # database functions are more complex to set up, so disable them by default
           use_database_authentication_functions? false