+++ test/rails_app/app/lib/rodauth_app.rb
@@ -40,10 +40,9 @@ class RodauthApp < Rodauth::Rails::App
     prefix "/admin"
   end
 
-  configure(:json, json: true) do
+  configure(:json) do
     enable :jwt, :create_account, :verify_account
     only_json? true
-    rails_controller { ActionController::API } if ActionPack.version >= Gem::Version.new("5.0")
     prefix "/json"
     jwt_secret "secret"
     account_status_column :status